feat: /auth/sign-in을 쿼리 캐시 초기화 지점으로 설정
