+++ apps/web/src/features/auth/lib/index.ts
@@ -1,2 +1,3 @@
 export { routeAfterSignIn } from './routeAfterSignIn';
+export { resetAuthAndRouteToSignIn } from './resetAuthAndRouteToSignIn';
 export { useKakaoSDK } from './useKakaoSDK';
