--- apps/web/src/shared/constants/auth/queryParams.ts
@@ -1,2 +1,0 @@
-export const CLEAR_QUERY_PARAM = 'clearQuery';
-export const CLEAR_QUERY_PARAM_VALUE = '1';
