+++ apps/web/src/shared/constants/index.ts
@@ -6,7 +6,6 @@ export * from './copy';
 export * from './routes';
 export * from './quick-menu';
 export * from './auth/publicEndpoints';
-export * from './auth/queryParams';
 export * from './form';
 export * from './urls';
 export * from './contact';
