--- apps/web/src/shared/ui/provider/clear-query/index.ts
@@ -1,1 +1,0 @@
-export { ClearQueryProvider } from './ClearQueryProvider';
