+++ apps/web/src/shared/ui/provider/index.ts
@@ -2,6 +2,6 @@ export {
   QueryErrorBoundary,
   QueryProviderWithDevtools,
 } from './tanstack-query';
-export { ClearQueryProvider } from './clear-query';
+export { QueryClientClearProvider } from './query-client-clear';
 export { HydrationSuspense } from './suspense/HydrationSuspense';
 export { Toaster } from './toast';
