+++ apps/web/src/shared/ui/provider/query-client-clear/index.ts
@@ -1,0 +1,1 @@
+export { QueryClientClearProvider } from './QueryClientClearProvider';
